requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn==0.27.1

# Database
supabase==1.2.0

# Validation and settings
pydantic>=2.4.2
pydantic[email]==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# HTTP and requests
requests==2.31.0

# Additional packages
python-dotenv==1.0.0
twilio==8.10.0
python-multipart==0.0.9  # For form data handling
